requirements.txt
yfinance
numpy
pandas
fastapi
langchain
pydantic
python-dotenv
mcp